\name{checkDuplicates_vect}
\alias{checkDuplicates_vect}
\title{
Check for duplicates
}
\description{
Check if an input vector has values that occur more than once. If so, return the frequency table
\usage{
checkDuplicates_vect(vect)
}
\arguments{
  \item{vect}{
Vector that contains numbers
}
}
\author{
Peter I-Fan Wu
}
\examples{
my_vector <- c(1,1,2,2,3)
checkDuplicates_vect(my_vector)
}

