\name{get_confusionMatrix_and_metrics}
\alias{get_confusionMatrix_and_metrics}
\title{
Compute confusion-matrix-related values and metrics
}
\description{
Generate a dataframe that contains: TP, FP, TN, FN, sensitivity, specificity, precision, accuracy for every pair as a cutoff and the same metrics based on randomization as negative controls.
}
\usage{
get_confusionMatrix_and_metrics(df, annot, similarity, seed = 103)
}
\arguments{
  \item{df}{
A table generated by generate_pairs_similarity_coannotation()
}
  \item{annot}{
The name of the column that contains whether pairs of genes are co-annotated or not
}
  \item{similarity}{
The name of the column that contains similarity scores, or an otherwise provided similarity vector
}
  \item{seed}{
The random seed number
}
}
\author{
Peter I-Fan Wu
}

