\name{microbialPhenotypes-package}
\alias{microbialPhenotypes-package}
\alias{microbialPhenotypes}
\docType{package}
\title{
\packageTitle{microbialPhenotypes}
}
\description{
\packageDescription{microbialPhenotypes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{microbialPhenotypes}
\packageIndices{microbialPhenotypes}
}
\author{
\packageAuthor{microbialPhenotypes}

Maintainer: \packageMaintainer{microbialPhenotypes}
}

\keyword{ package }

